refactor: consolidating duplicate axis code
