+++ packages/vchart/src/component/axis/cartesian/axis.ts
@@ -439,54 +439,20 @@ export abstract class CartesianAxis<T extends ICartesianAxisCommonSpec = ICartes
     return scales;
   }
 
-  protected collectData(depth?: number, rawData?: boolean) {
-    const data: { min: number; max: number; values: any[] }[] = [];
-    eachSeries(
-      this._regions,
-      s => {
-        let field: string | string[];
-        if (depth > 0) {
-          field = s.getGroups()?.fields?.[depth];
-        } else {
-          if (isXAxis(this.getOrient())) {
-            field = (s as ICartesianSeries).getSpec().x2Field
-              ? [...(s as ICartesianSeries).fieldX, (s as ICartesianSeries).fieldX2]
-              : (s as ICartesianSeries).fieldX;
-          } else if (isZAxis(this.getOrient())) {
-            field = (s as ICartesianSeries).fieldZ;
-          } else {
-            field = (s as ICartesianSeries).getSpec().y2Field
-              ? [...(s as ICartesianSeries).fieldY, (s as ICartesianSeries).fieldY2]
-              : (s as ICartesianSeries).fieldY;
-          }
-        }
-        field = (isArray(field) ? (isContinuous(this._scale.type) ? field : [field[0]]) : [field]) as string[];
-        if (!depth) {
-          this._dataFieldText = s.getFieldAlias(field[0]);
-        }
-        if (field) {
-          const viewData = s.getViewData();
-          if (rawData) {
-            field.forEach(f => {
-              data.push(s.getRawDataStatisticsByField(f, false) as { min: number; max: number; values: any[] });
-            });
-          } else if (viewData && viewData.latestData && viewData.latestData.length) {
-            const seriesData = s.getViewDataStatistics?.();
-
-            field.forEach(f => {
-              if (seriesData?.latestData?.[f]) {
-                data.push(seriesData.latestData[f]);
-              }
-            });
-          }
-        }
-      },
-      {
-        userId: this._seriesUserId,
-        specIndex: this._seriesIndex
+  protected collectSeriesField(depth: number, series: ICartesianSeries) {
+    let field: string | string[];
+    if (depth > 0) {
+      field = series.getGroups()?.fields?.[depth];
+    } else {
+      if (isXAxis(this.getOrient())) {
+        field = series.getSpec().x2Field ? [...series.fieldX, series.fieldX2] : series.fieldX;
+      } else if (isZAxis(this.getOrient())) {
+        field = series.fieldZ;
+      } else {
+        field = series.getSpec().y2Field ? [...series.fieldY, series.fieldY2] : series.fieldY;
       }
-    );
-    return data;
+    }
+    return field;
   }
 
   protected updateSeriesScale() {
